Refactor TerminatorInsn class
